check_snmp: Remove that is_numeric madness
Original patch to make Timeticks works as in check_snmp v1.4.14, it turns
out is_numeric isn't so useful and treating all types as numeric works
best for backwards-compatibility. This is how it used to work in 1.4.14.

As a special case, I also make calculate_rate look up for numeric values
as it would otherwise return the last value instead.

This patch cherry-picks multiple patches from the master branch:
  896962a1ad1b7d7c75d42c565b06cc799feb0a7c check_snmp now considers strings returned by SNMP that contain just numbers (according to strtod) to be a numeric value
  df88f95fcaf65d58a9ea172c2b3e2b96d80dff33 check_snmp: Remove that is_numeric madness
  811684ffe394d050158c2a98d5be211d5ded3c88 State-based tests enhancements
  5a2814a21bff07b87a7589ef19b63c9eecb8be9f Revert "check_snmp now considers strings returned by SNMP that contain just"